refactor(Link): move the prefetch wiring into Link.vue
